requirements.txt
langchain
langchain-xai
python-dotenv
langchain-pinecone
pinecone-client
gradio
langchain-core
langchain-openai